:bug: fix: tipo do campo chat_id de solicitacao_agenda errado

+++ script_eleve.sql
@@ -82,7 +82,7 @@ CREATE TABLE agenda_servico (
 
 CREATE TABLE solicitacao_agenda (
     id INT PRIMARY KEY AUTO_INCREMENT,
-    chat_id INT NOT NULL,
+    chat_id BIGINT NOT NULL,
     pet_id INT NOT NULL,
     valor_deslocamento FLOAT,
     data_hora_inicio DATETIME NOT NULL,
